requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
redis==5.0.4
pg8000==1.31.1
httpx==0.27.0

